tests/wpt/web-platform-tests/webrtc/RTCPeerConnection-setRemoteDescription-offer.html

<!doctype html>
<meta charset=utf-8>
<title>RTCPeerConnection.prototype.setRemoteDescription - offer</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="RTCPeerConnection-helper.js"></script>
<script>
  'use strict';

  // Test is based on the following editor draft:
  // https://w3c.github.io/webrtc-pc/archives/20170605/webrtc.html

  // The following helper functions are called from RTCPeerConnection-helper.js:
  //   generateOffer()
  //   assert_session_desc_equals()
  //   test_state_change_event()

  /*
    4.3.2.  Interface Definition
      [Constructor(optional RTCConfiguration configuration)]
      interface RTCPeerConnection : EventTarget {
        Promise<void>                      setRemoteDescription(
            RTCSessionDescriptionInit description);

        readonly attribute RTCSessionDescription? remoteDescription;
        readonly attribute RTCSessionDescription? currentRemoteDescription;
        readonly attribute RTCSessionDescription? pendingRemoteDescription;
        ...
      };

    4.6.2.  RTCSessionDescription Class
      dictionary RTCSessionDescriptionInit {
        required RTCSdpType type;
                 DOMString  sdp = "";
      };

    4.6.1.  RTCSdpType
      enum RTCSdpType {
        "offer",
        "pranswer",
        "answer",
        "rollback"
      };
   */

  /*
    4.3.1.6.  Set the RTCSessionSessionDescription
      2.2.3.  Otherwise, if description is set as a remote description, then run one of
              the following steps:
        - If description is of type "offer", set connection.pendingRemoteDescription
          attribute to description and signaling state to have-remote-offer.
   */
  promise_test(t => {
    const pc = new RTCPeerConnection();

    test_state_change_event(t, pc, ['have-remote-offer']);

    return generateOffer({ pc, data: true })
    .then(offer =>
      pc.setRemoteDescription(offer)
      .then(() => {
        assert_equals(pc.signalingState, 'have-remote-offer');
        assert_session_desc_equals(pc.remoteDescription, offer);
        assert_session_desc_equals(pc.pendingRemoteDescription, offer);
        assert_equals(pc.currentRemoteDescription, null);
      }));
  }, 'setRemoteDescription with valid offer should succeed');

  promise_test(t => {
    const pc = new RTCPeerConnection();

    // have-remote-offer event should only fire once
    test_state_change_event(t, pc, ['have-remote-offer']);

    return Promise.all([
      pc.createOffer({ offerToReceiveAudio: true }),
      generateOffer({ pc, data: true })
    ]).then(([offer1, offer2]) =>
      pc.setRemoteDescription(offer1)
      .then(() => pc.setRemoteDescription(offer2))
      .then(() => {
        assert_equals(pc.signalingState, 'have-remote-offer');
        assert_session_desc_equals(pc.remoteDescription, offer2);
        assert_session_desc_equals(pc.pendingRemoteDescription, offer2);
        assert_equals(pc.currentRemoteDescription, null);
      }));
  }, 'Setting remote description multiple times with different offer should succeed');

  /*
    4.3.1.6.  Set the RTCSessionSessionDescription
      2.1.4.  If the content of description is not valid SDP syntax, then reject p with
              an RTCError (with errorDetail set to "sdp-syntax-error" and the
              sdpLineNumber attribute set to the line number in the SDP where the syntax
              error was detected) and abort these steps.
   */
  promise_test(t => {
    const pc = new RTCPeerConnection();

    return pc.setRemoteDescription({
      type: 'offer',
      sdp: 'Invalid SDP'
    })
    .then(() => {
      assert_unreached('Expect promise to be rejected');
    }, err => {
      assert_equals(err.errorDetail, 'sdp-syntax-error',
        'Expect error detail field to set to sdp-syntax-error');

      assert_true(err instanceof RTCError,
        'Expect err to be instance of RTCError');
    });
  }, 'setRemoteDescription(offer) with invalid SDP should reject with RTCError');

  /*
    4.3.1.6.  Set the RTCSessionSessionDescription
      2.1.3.  If the description's type is invalid for the current signaling state of
              connection, then reject p with a newly created InvalidStateError and abort
              these steps.

    [JSEP]
      5.6.  If the type is "offer", the PeerConnection state MUST be either "stable" or
            "have-remote-offer".
   */
  promise_test(t => {
    const pc = new RTCPeerConnection();

    return pc.createOffer()
    .then(offer => pc.setLocalDescription(offer))
    .then(() => pc.createOffer())
    .then(offer2 =>
      promise_rejects(t, 'InvalidStateError',
        pc.setRemoteDescription(offer2)));
  }, 'Calling setRemoteDescription(offer) from have-local-offer state should reject with InvalidStateError');

</script>
